代理商功能联调修改

+++ src/main/java/com/bdaim/rbac/service/UserService.java
@@ -1883,12 +1883,12 @@ public class UserService {
                 cell = row.createCell(c++);
                 cell.setCellValue((String) map.get("commision"));
                 cell = row.createCell(c++);
-                cell.setCellValue(((BigDecimal)map.get("dataAmcount")).toString());
+                cell.setCellValue(((Double)map.get("dataAmcount")).toString());
                 cell = row.createCell(c++);
-                cell.setCellValue(((BigDecimal) map.get("callAmcount")).toString());
+                cell.setCellValue(((Double) map.get("callAmcount")).toString());
                 cell = row.createCell(c++);
                 ((BigDecimal)map.get("dataAmcount")).add(((BigDecimal)map.get("callAmcount")).add((BigDecimal) map.get("callAmcount")));
-                cell.setCellValue(((BigDecimal) map.get("messageAmcount")).toString());
+                cell.setCellValue(((Double) map.get("messageAmcount")).toString());
                 cell = row.createCell(c++);
                 cell.setCellValue(((BigDecimal)map.get("dataAmcount")).add(((BigDecimal)map.get("callAmcount")).add((BigDecimal) map.get("callAmcount"))).toString());
             }
